Extract appimage_hexlify into separate source file
