+++ src/appimagetool_shared.c
@@ -187,18 +187,3 @@ bool appimage_type2_digest_md5(const char* path, char* digest) {
 
     return true;
 }
-
-char* appimage_hexlify(const char* bytes, const size_t numBytes) {
-    // first of all, allocate the new string
-    // a hexadecimal representation works like "every byte will be represented by two chars"
-    // additionally, we need to null-terminate the string
-    char* hexlified = (char*) calloc((2 * numBytes + 1), sizeof(char));
-
-    for (size_t i = 0; i < numBytes; i++) {
-        char buffer[3];
-        sprintf(buffer, "%02x", (unsigned char) bytes[i]);
-        strcat(hexlified, buffer);
-    }
-
-    return hexlified;
-}
