+++ tests/CMakeLists.txt
@@ -29,6 +29,7 @@ if(BUILD_TESTING)
         ${PROJECT_SOURCE_DIR}/src/elf.c
         ${PROJECT_SOURCE_DIR}/src/getsection.c
         ${PROJECT_SOURCE_DIR}/src/appimagetool_shared.c
+        ${PROJECT_SOURCE_DIR}/src/hexlify.c
     )
 
     target_link_libraries(test_shared
